+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * Streams of integers from a binomial distribution.
 * Streams of doubles from an exponential distribution.
+* Streams of doubles from a Gaussian distribution.
 * Streams of not strictly uniform random integers.
 * Seeded constructors for the EnhancedSplittableGenerator, EnhancedStreamableGenerator, and
   EnhancedRandomGenerator classes.